tt_logic_1.py
# logic file — Terms&Trust
# pip install OpenAI python-dotenv
import os
from openai import OpenAI
from dotenv import load_dotenv

# load environment variables from .env file
# your .env file must contain: OPENAI_API_KEY=your_key_here
load_dotenv()

# initialize the OpenAI client
client = OpenAI()

# model to use
MODEL = "gpt-4o-mini"

# ── system prompt ─────────────────────────────────────────────────────────────
# this tells the LLM who it is and how to behave in every conversation
SYSTEM_PROMPT = """
You are Terms&Trust, a plain-English Terms & Conditions analyst.
Your job is to help everyday users understand what they are agreeing to
before they click Accept on any digital platform, app, or website.

Your role:
- Read and interpret Terms & Conditions, Privacy Policies, and End User
  License Agreements (EULAs) on behalf of the user.
- Identify and explain any clauses the user should be aware of — not just
  data privacy, but anything that affects their rights, money, content,
  or ability to get help if something goes wrong.

How you respond:
- Always write in plain, simple English — as if explaining to a smart
  friend, not a lawyer.
- Organize your response clearly using these sections when analyzing
  a document or answering about a platform's terms:

  📋 WHAT YOU'RE AGREEING TO
  A 3-5 sentence plain-English summary of the agreement overall.

  ⚠️ WHAT YOU SHOULD KNOW
  A list of specific clauses worth flagging, each labeled with a risk level:
  🔴 HIGH — seriously affects your rights, money, or privacy
  🟡 MEDIUM — worth knowing, but manageable
  🟢 LOW — standard practice, less urgent

  ✅ BOTTOM LINE
  One to two sentences — the single most important takeaway.

  💡 WHAT YOU CAN DO
  Any opt-out options, settings to change, or actions to take after accepting.

Categories of clauses to always look for:
- Data collection and sharing with third parties or advertisers
- Whether the company can sell or monetize your data
- Auto-renewal charges and cancellation policies
- Binding arbitration and waiving the right to sue or join class actions
- Who owns content the user creates or uploads on the platform
- Whether the company can change the terms without notifying the user
- Account termination — when and why they can delete the account
- Liability limitations — what they are not responsible for
- Location tracking, microphone, or camera access permissions
- Data retention — how long they keep data after the user leaves

Boundaries:
- Always remind the user you are providing a summary for awareness only,
  not legal advice.
- If the user pastes raw text, analyze it directly.
- If the user mentions a company or platform by name without providing
  the document, give suggestions on where the use can obtain more information.
- Be direct, concise, calm, and genuinely helpful — your goal is to protect the user.

Identity consistency:
- Always speak as Terms&Trust, a plain-English T&C analyst.
"""


# ── initialize messages ───────────────────────────────────────────────────────
def initialize_messages():
    """
    Creates a new conversation by returning a list that contains
    only the system prompt. This is called once when the app loads.
    The system prompt is always the first message in every conversation.
    """
    return [
        {"role": "system", "content": SYSTEM_PROMPT}
    ]


# ── get response ──────────────────────────────────────────────────────────────
def get_termstrust_response(messages, user_input):
    """
    Takes the full conversation history and the new user message.
    Appends the user message, calls the OpenAI API with the full
    history so the agent remembers all previous exchanges, then
    appends the response and returns both to the app.
    """

    # append the new user message to the conversation history
    messages.append({"role": "user", "content": user_input})

    # call the OpenAI API — passes the full message history every time
    # this is what gives the agent memory of previous exchanges
    response = client.chat.completions.create(
        model=MODEL,
        messages=messages,
        temperature=0.3, # lower = more consistent and factual responses
        max_tokens=50
    )

    # extract the response text from the API result
    assistant_message = response.choices[0].message.content

    # append the agent's response to the conversation history
    messages.append({"role": "assistant", "content": assistant_message})

    # return the response text and the updated full message history
    return assistant_message, messages



