requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
requests
beautifulsoup4
pdfplumber
python-dotenv
pypdf
docx2txt
